LayoutTests/imported/w3c/web-platform-tests/css/css-scrollbars/scrollbar-color-003-expected.txt

Test scrollbar-color: vertical scrollbar
auto
themed

PASS scrollbar-color auto defers to ::-webkit-scrollbar
FAIL scrollbar-color yellow blue overrides ::-webkit-scrollbar assert_less_than: themed scrollWidth expected a number less than 200 but got 200
